remove unnecessary document load from outline and move it to viewer
