--- projects/media-viewer/src/lib/toolbar/side-bar/outline-view/outline-item/outline-item.component.html
@@ -1,14 +1,0 @@
-<div #outlineItem class="outlineItem">
-  <div *ngIf="outline.items.length > 0" [ngClass]="{ 'outlineItemToggler': true, 'outlineItemsHidden': !showOutlineItems }"
-       (click)="toggleOutline()" (keyup.enter)="toggleOutline()" tabindex="0"></div>
-  <a (click)="navigateLink()" (keyup.enter)="navigateLink()" [ngStyle]="getStyles()" tabindex="0">
-    {{ outline.title }}
-  </a>
-  <div *ngIf="outline.items.length > 0" class="outlineItems">
-    <div *ngFor="let outline of outline.items">
-      <mv-outline-item
-        [outline]="outline"></mv-outline-item>
-    </div>
-  </div>
-</div>
-
