--- projects/media-viewer/src/lib/toolbar/side-bar/outline-view/outline-view.component.html
@@ -1,8 +1,0 @@
-<div #outlineContainer id="outlineView" class="outlineWithDeepNesting">
-  <div class="outlineItem">
-    <ng-container *ngFor="let outline of outline">
-      <mv-outline-item
-        [outline]="outline"></mv-outline-item>
-    </ng-container>
-  </div>
-</div>
