--- projects/media-viewer/src/lib/toolbar/side-bar/side-bar.component.html
@@ -1,7 +1,0 @@
-<div>
-  <div id="sidebarContent">
-    <mv-outline-view [url]="url">
-    </mv-outline-view>
-  </div>
-  <div id="sidebarResizer"></div>
-</div>
