--- projects/media-viewer/src/lib/toolbar/side-bar/side-bar.component.ts
@@ -1,13 +1,0 @@
-import { Component, Input } from '@angular/core';
-
-@Component({
-  selector: 'mv-side-bar',
-  templateUrl: './side-bar.component.html',
-})
-export class SideBarComponent {
-
-  @Input() url: string;
-
-  constructor() {
-  }
-}
