+++ projects/media-viewer/src/lib/viewers/pdf-viewer/outline-view/outline-view.component.html
@@ -1,0 +1,10 @@
+<div id="sidebarContent">
+  <div id="outlineView" class="outlineWithDeepNesting">
+    <div class="outlineItem">
+      <mv-outline-item *ngFor="let outlineItem of outline"
+                       [outline]="outlineItem"></mv-outline-item>
+    </div>
+  </div>
+</div>
+<div id="sidebarResizer"></div>
+
