+++ projects/media-viewer/src/lib/viewers/pdf-viewer/outline-view/outline-view.component.spec.ts
@@ -1,0 +1,28 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { OutlineViewComponent } from './outline-view.component';
+import { OutlineItemComponent } from './outline-item/outline-item.component';
+
+describe('OutlineViewComponent', () => {
+  let component: OutlineViewComponent;
+  let fixture: ComponentFixture<OutlineViewComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [ OutlineViewComponent, OutlineItemComponent ],
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(OutlineViewComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+
+
+});
