+++ projects/media-viewer/src/lib/viewers/pdf-viewer/outline-view/outline-view.component.ts
@@ -1,0 +1,11 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'mv-outline-view',
+  templateUrl: './outline-view.component.html',
+  styleUrls: ['../../../styles/main.scss'],
+})
+export class OutlineViewComponent {
+
+  @Input() outline: [];
+}
