+++ projects/media-viewer/src/lib/viewers/pdf-viewer/pdf-viewer.component.html
@@ -1,3 +1,7 @@
+<mv-outline-view *ngIf="toolbarButtons.sidebarOpen"
+                 id="sidebarContainer"
+                 [outline]="documentOutline">
+</mv-outline-view>
 <mv-comment-set-toggle *ngIf="enableAnnotations && annotationSet"></mv-comment-set-toggle>
 
 <mv-comment-set-header [ngClass]="{'show-comments-panel': showCommentsPanel}"
